+++ src/Extension/TableOfContents/TableOfContentsBuilder.php
@@ -41,7 +41,8 @@ final class TableOfContentsBuilder implements ConfigurationAwareInterface
             (string) $this->config->get('table_of_contents/style'),
             (string) $this->config->get('table_of_contents/normalize'),
             (int) $this->config->get('table_of_contents/min_heading_level'),
-            (int) $this->config->get('table_of_contents/max_heading_level')
+            (int) $this->config->get('table_of_contents/max_heading_level'),
+            (string) $this->config->get('heading_permalink/fragment_prefix'),
         );
 
         $toc = $generator->generate($document);